Extract RecordDocumentReference from GroundFirestore
